refactor: html 태그와 혼동되는 컴포넌트 이름 변경
