+++ frontend/src/components/DictInfo/index.tsx
@@ -1,5 +1,5 @@
 import Title from './ingredients/Title.style';
-import Content from './ingredients/Content';
+import Content from '../Tag';
 import Main from './ingredients/Main';
 
 const DictInfo = Object.assign(Main, { Title, Content });
